Add parameter to instruct refresh workers not to refresh for when cached result is empty
Empty results (as cache in DB) trigger a onZeroLoaded callback from the paging library
and an inital query request by itself.
Running the worker in this situation leads to two queries being send out.
This commit adds a parameter to the workers that lets them skip (not refresh) when
the current results are empty.

The alternative would be to not let the paging library load on empty results.
However that callback also triggers some nice loading indicators that we want
to keep