EventMonitorService should not start listening if it was destroyed
On >= Android 8 START_STICKY will result in the EventMonitorService getting
destroyed and recreated when the app is swiped away. Only to be destroyed
seconds afterwards again. This results in the onAccountsLoaded() callback
being fired after the service has been destroyed. We need to detect this
to not register callbacks with the PushService otherwise it will leak the
service.