introduce BuildConfig flag to use Foreground service
The EventMonitorService has been designed as a  background service while the
app is running (to have a central point that listens for push events and
invokes Query refreshes)

However the very same design can be used to permanently listen for events.

The only difference is whether or not the EventMonitorService spawns a
foreground notification.

A build flag will enable that going forward. Release versions will probably
not use that but it is a nice alternative for testing.