disable selection tracking during swipe
Previously a swipe (with resting in a position for a moment) might have
triggered a long press that selected the item.

On one hand this could cause race conditons where the swipe was first completed
internally and then the selection tracker didn’t find the object in question any
more.

(fixes #77)

on the other hand the behaviour (selecting while swiping) was just weird and
unexpected UX.