create custom selection tracker
The SelectionTracker from AndroidX handles its own MotionEvents. This causes
some bad interactions with swipe support. For example it would trigger a long
press after the item had already been swiped. Causing the next item in the list
to be selected.

This commit introduces a custom selection tracker that is backed by a simple
Set<String> in the ViewModel.
Some of it's api has been kept the same to make the changeset small