show request help button in session resource screen
Ltt.rs generally tries to make the onboarding screens as clean as possible
and not overload them with too much information. However in the session
resource / connection URL screen we need to make a small exception as JMAP
just isn’t as popular yet.

Besides most users will be able to skip this screen anyway as the URL is
discoverd automatically.

fixes #35