fix regression that showed toolbar title in thread view
A previous commit introduced a regression that caused the thread view to show
the title from the previous fragment.

Previously the toolbar title was hidden when navigating to the thread fragment.

This commit let's ThreadFragment explicitly set the title to null instead of
hiding it. (This also cleans up the code in that thread desination doesn’t
get special threatment anymore.)