DiffUtil.ItemCallback<> needs working equals() to not redraw
The objects we put into RecyclerView need working equals() and hashCode()
otherwise they will always be detected as changed and trigger a redraw.

This leads to jerky UI (for example when animating an insert).

This commit fixes a regression introduced after reworking the From class which
didn’t have a proper equals.